% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{2 sample test
This is my constructor function. This function will create a list. The constructor runs the hypotheses on two samples, and creates and object with the confidence interval and p statistics}
\usage{
myconstr(x, y, alpha)
}
\arguments{
\item{x}{sample 1 data}

\item{y}{sample 2 data}

\item{alpha}{significance level}
}
\value{
list of data, confidence interval and p value
}
\description{
2 sample test
This is my constructor function. This function will create a list. The constructor runs the hypotheses on two samples, and creates and object with the confidence interval and p statistics
}
\examples{
\dontrun{obj = myconstr(x = x, y = y, alpha = 0.05)}
}
