% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{plot
Generates a boxplot of two samples}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{Rttest object}

\item{...}{other}
}
\value{
boxplot of two samples
}
\description{
plot
Generates a boxplot of two samples
}
\examples{
\dontrun{plot(obj)}

}
