% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Prints out two samples, confidence interval and significance level to a table.}
\usage{
\method{print}{Rttest}(x)
}
\arguments{
\item{x}{Rttest object}
}
\value{
a table with two samples, confidence interval and significance level
}
\description{
Prints out two samples, confidence interval and significance level to a table.
}
\examples{
\dontrun{print(obj)}
}
