% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{print
Prints out two samples, confidence interval and significance level to a table.}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{Rttest object}

\item{...}{other}
}
\value{
a table with two samples, confidence interval and significance level
}
\description{
print
Prints out two samples, confidence interval and significance level to a table.
}
\examples{
\dontrun{print(obj)}

}
